+++ src/mplchart/stylesheet.py
@@ -7,11 +7,4 @@ chart:bgcolor   Chart background color
 
 """
 
-STYLESHEET = {
-    "chart:bgcolor": None,
-    "candlesticks.up:color": "black",
-    "candlesticks.down:color": "red",
-    "line:color": "black",
-    "sma:color": ["blue", "green", "red"],
-    "rsi:color": "black",
-}
+STYLESHEET = {}